+++ src/main/java/com/verdantartifice/primalmagic/common/theorycrafting/AbstractProjectMaterial.java
@@ -1,5 +1,6 @@
 package com.verdantartifice.primalmagic.common.theorycrafting;
 
+import java.util.Objects;
 import java.util.Set;
 
 import javax.annotation.Nonnull;
@@ -9,6 +10,7 @@ import com.verdantartifice.primalmagic.common.research.CompoundResearchKey;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.util.INBTSerializable;
@@ -21,6 +23,7 @@ import net.minecraftforge.common.util.INBTSerializable;
 public abstract class AbstractProjectMaterial implements INBTSerializable<CompoundTag> {
     protected boolean selected;
     protected double weight;
+    protected double bonusReward;
     protected CompoundResearchKey requiredResearch;
     
     protected AbstractProjectMaterial() {
@@ -33,6 +36,7 @@ public abstract class AbstractProjectMaterial implements INBTSerializable<Compou
         retVal.putString("MaterialType", this.getMaterialType());
         retVal.putBoolean("Selected", this.isSelected());
         retVal.putDouble("Weight", this.getWeight());
+        retVal.putDouble("BonusReward", this.bonusReward);
         if (this.requiredResearch != null) {
             retVal.putString("RequiredResearch", this.getRequiredResearch().toString());
         }
@@ -43,6 +47,7 @@ public abstract class AbstractProjectMaterial implements INBTSerializable<Compou
     public void deserializeNBT(CompoundTag nbt) {
         this.selected = nbt.getBoolean("Selected");
         this.weight = nbt.getDouble("Weight");
+        this.bonusReward = nbt.getDouble("BonusReward");
         this.requiredResearch = nbt.contains("RequiredResearch") ? CompoundResearchKey.parse(nbt.getString("RequiredResearch")) : null;
     }
     
@@ -92,6 +97,10 @@ public abstract class AbstractProjectMaterial implements INBTSerializable<Compou
         return this.weight;
     }
     
+    public double getBonusReward() {
+        return this.bonusReward;
+    }
+    
     @Nullable
     public CompoundResearchKey getRequiredResearch() {
         return this.requiredResearch;
@@ -105,8 +114,16 @@ public abstract class AbstractProjectMaterial implements INBTSerializable<Compou
         this.weight = weight;
     }
     
+    public void setBonusReward(double bonus) {
+        this.bonusReward = bonus;
+    }
+    
     public void setRequiredResearch(@Nonnull CompoundResearchKey key) {
         this.requiredResearch = key.copy();
+    }
+    
+    public boolean isAllowedInProject(ServerPlayer player) {
+        return this.hasRequiredResearch(player);
     }
     
     public boolean hasRequiredResearch(Player player) {
@@ -121,14 +138,7 @@ public abstract class AbstractProjectMaterial implements INBTSerializable<Compou
 
     @Override
     public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((requiredResearch == null) ? 0 : requiredResearch.hashCode());
-        result = prime * result + (selected ? 1231 : 1237);
-        long temp;
-        temp = Double.doubleToLongBits(weight);
-        result = prime * result + (int) (temp ^ (temp >>> 32));
-        return result;
+        return Objects.hash(bonusReward, requiredResearch, selected, weight);
     }
 
     @Override
@@ -140,15 +150,8 @@ public abstract class AbstractProjectMaterial implements INBTSerializable<Compou
         if (getClass() != obj.getClass())
             return false;
         AbstractProjectMaterial other = (AbstractProjectMaterial) obj;
-        if (requiredResearch == null) {
-            if (other.requiredResearch != null)
-                return false;
-        } else if (!requiredResearch.equals(other.requiredResearch))
-            return false;
-        if (selected != other.selected)
-            return false;
-        if (Double.doubleToLongBits(weight) != Double.doubleToLongBits(other.weight))
-            return false;
-        return true;
+        return Double.doubleToLongBits(bonusReward) == Double.doubleToLongBits(other.bonusReward)
+                && Objects.equals(requiredResearch, other.requiredResearch) && selected == other.selected
+                && Double.doubleToLongBits(weight) == Double.doubleToLongBits(other.weight);
     }
 }