+++ src/main/java/com/verdantartifice/primalmagic/common/theorycrafting/ProjectTemplate.java
@@ -11,7 +11,6 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSyntaxException;
-import com.verdantartifice.primalmagic.common.capabilities.IPlayerKnowledge.KnowledgeType;
 import com.verdantartifice.primalmagic.common.research.SimpleResearchKey;
 import com.verdantartifice.primalmagic.common.stats.StatsManager;
 import com.verdantartifice.primalmagic.common.stats.StatsPM;
@@ -19,6 +18,7 @@ import com.verdantartifice.primalmagic.common.util.WeightedRandomBag;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.registries.ForgeRegistries;
 
@@ -58,7 +58,7 @@ public class ProjectTemplate {
     }
     
     @Nullable
-    public Project initialize(Player player) {
+    public Project initialize(ServerPlayer player) {
         if (this.requiredResearch != null && !this.requiredResearch.isKnownByStrict(player)) {
             // Fail initialization to prevent use if the player doesn't have the right research unlocked
             return null;
@@ -78,7 +78,7 @@ public class ProjectTemplate {
         }
         
         // Create new initialized project
-        return new Project(this.key, materials, this.getBaseSuccessChance(player), this.getTheoryPointReward(), this.aidBlock);
+        return new Project(this.key, materials, this.getBaseSuccessChance(player), this.rewardMultiplier, this.aidBlock);
     }
     
     protected int getRequiredMaterialCount(Player player) {
@@ -98,18 +98,14 @@ public class ProjectTemplate {
     }
     
     @Nonnull
-    protected WeightedRandomBag<AbstractProjectMaterial> getMaterialOptions(Player player) {
+    protected WeightedRandomBag<AbstractProjectMaterial> getMaterialOptions(ServerPlayer player) {
         WeightedRandomBag<AbstractProjectMaterial> retVal = new WeightedRandomBag<>();
         for (AbstractProjectMaterial material : this.materialOptions) {
-            if (material.hasRequiredResearch(player)) {
+            if (material.isAllowedInProject(player)) {
                 retVal.add(material, material.getWeight());
             }
         }
         return retVal;
-    }
-    
-    protected int getTheoryPointReward() {
-        return (int)(KnowledgeType.THEORY.getProgression() * this.rewardMultiplier);
     }
     
     public static class Serializer implements IProjectTemplateSerializer {