+++ src/main/java/com/verdantartifice/primalmagic/datagen/theorycrafting/ExperienceMaterialBuilder.java
@@ -11,6 +11,7 @@ public class ExperienceMaterialBuilder {
     protected final int levels;
     protected final boolean consumed;
     protected double weight = 1D;
+    protected double bonusReward = 0.0D;
     protected CompoundResearchKey requiredResearch;
 
     protected ExperienceMaterialBuilder(int levels, boolean consumed) {
@@ -40,6 +41,11 @@ public class ExperienceMaterialBuilder {
         return requiredResearch(CompoundResearchKey.parse(keyStr));
     }
     
+    public ExperienceMaterialBuilder bonusReward(double bonus) {
+        this.bonusReward = bonus;
+        return this;
+    }
+    
     private void validate() {
         if (this.levels <= 0) {
             throw new IllegalStateException("Invalid levels for experience project material");
@@ -47,23 +53,28 @@ public class ExperienceMaterialBuilder {
         if (this.weight <= 0D) {
             throw new IllegalStateException("Invalid weight for experience project material");
         }
+        if (this.bonusReward < 0.0D) {
+            throw new IllegalStateException("Invalid bonus reward for experience project material");
+        }
     }
     
     public IFinishedProjectMaterial build() {
         this.validate();
-        return new ExperienceMaterialBuilder.Result(this.levels, this.consumed, this.weight, this.requiredResearch);
+        return new ExperienceMaterialBuilder.Result(this.levels, this.consumed, this.weight, this.bonusReward, this.requiredResearch);
     }
     
     public static class Result implements IFinishedProjectMaterial {
         protected final int levels;
         protected final boolean consumed;
         protected final double weight;
+        protected final double bonusReward;
         protected final CompoundResearchKey requiredResearch;
 
-        public Result(int levels, boolean consumed, double weight, @Nullable CompoundResearchKey requiredResearch) {
+        public Result(int levels, boolean consumed, double weight, double bonusReward, @Nullable CompoundResearchKey requiredResearch) {
             this.levels = levels;
             this.consumed = consumed;
             this.weight = weight;
+            this.bonusReward = bonusReward;
             this.requiredResearch = requiredResearch == null ? null : requiredResearch.copy();
         }
 
@@ -73,6 +84,9 @@ public class ExperienceMaterialBuilder {
             json.addProperty("levels", this.levels);
             json.addProperty("consumed", this.consumed);
             json.addProperty("weight", this.weight);
+            if (this.bonusReward > 0.0D) {
+                json.addProperty("bonus_reward", this.bonusReward);
+            }
             if (this.requiredResearch != null) {
                 json.addProperty("required_research", this.requiredResearch.toString());
             }